Refactor: Create separate files for ranklist functions
